Rename ConsumerProducer.cpp to Consumer/Multithreading.cpp
This is showing a consumer process receiving from a buffer and a producer process putting data into the buffer. Each checks and waits to see if there is anything in it or if it's full. They both wait accordingly. This is also an example of using an API and using multithreading.